+++ 2022/day11/src/Monkey.java
@@ -2,20 +2,20 @@ import java.util.List;
 
 public class Monkey {
     final int number;
-    final List<Integer> items;
+    final List<Long> items;
     final Operation opr;
     final int test, t, f;
 
     public int inspections = 0;
 
-    public Monkey(int number, List<Integer> items, String opr, int test, int t, int f) {
+    public Monkey(int number, List<Long> items, String opr, int test, int t, int f) {
         this.number = number;
         this.items = items;
 
         final char symbol = opr.charAt(0);
         String s = opr.split(" ")[1];
         this.opr = (n -> {
-            int operand = s.equals("old") ? n : Integer.parseInt(s);
+            long operand = s.equals("old") ? n : Integer.parseInt(s);
             return symbol == '+' ? n + operand : n * operand;
         });
 
@@ -24,12 +24,12 @@ public class Monkey {
         this.f = f;
     }
 
-    public int inspect() {
+    public long inspect() {
         inspections++;
         return opr.doOpr(items.remove(0));
     }
 
-    public int test(int worry) {
+    public int test(long worry) {
         return worry % test == 0 ? t : f;
     }
 
@@ -38,6 +38,6 @@ public class Monkey {
     }
 
     private interface Operation {
-        int doOpr(int n);
+        long doOpr(long n);
     }
 }
